requirements.txt
toml
PyNaCl
